+++ CHANGES.md
@@ -4,7 +4,9 @@
 2.3b5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Do not clobber git clones by default. Add new option `clobber` to allow
+  specifying that a git clone may clobber the target directory.
+  ([#298](https://github.com/flyingcircusio/batou/issues/298))
 
 
 2.3b4 (2022-08-22)